note: Add a list of ignored phrases
Some people have auto-responders set up for one reason or another. This
adds a config value of phrases you'd like to not count as the user
coming back. Since I have a very simple case in practice and this is run
for every message regexes didn't sound like a great idea.